Remove Regex suffix from new rule names
