+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/rule/codestyle/FormalParameterNamingConventionsRule.java
@@ -11,7 +11,7 @@ import java.util.regex.Pattern;
 import net.sourceforge.pmd.lang.apex.ast.ASTParameter;
 import net.sourceforge.pmd.properties.PropertyDescriptor;
 
-public class FormalParameterRegexNamingConventionsRule extends AbstractRegexNamingConventionsRule {
+public class FormalParameterNamingConventionsRule extends AbstractNamingConventionsRule {
     private static final Map<String, String> DESCRIPTOR_TO_DISPLAY_NAME = new HashMap<>();
 
     private static final PropertyDescriptor<Pattern> FINAL_METHOD_PARAMETER_REGEX = prop("finalMethodParameterPattern", "final method parameter",
@@ -20,9 +20,14 @@ public class FormalParameterRegexNamingConventionsRule extends AbstractRegexNami
     private static final PropertyDescriptor<Pattern> METHOD_PARAMETER_REGEX = prop("methodParameterPattern", "method parameter",
             DESCRIPTOR_TO_DISPLAY_NAME).defaultValue(CAMEL_CASE).build();
 
-    public FormalParameterRegexNamingConventionsRule() {
+    public FormalParameterNamingConventionsRule() {
         definePropertyDescriptor(FINAL_METHOD_PARAMETER_REGEX);
         definePropertyDescriptor(METHOD_PARAMETER_REGEX);
+
+        setProperty(CODECLIMATE_CATEGORIES, "Style");
+        // Note: x10 as Apex has not automatic refactoring
+        setProperty(CODECLIMATE_REMEDIATION_MULTIPLIER, 1);
+        setProperty(CODECLIMATE_BLOCK_HIGHLIGHTING, false);
 
         addRuleChainVisit(ASTParameter.class);
     }
