+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/rule/codestyle/MethodNamingConventionsRule.java
@@ -6,21 +6,38 @@ package net.sourceforge.pmd.lang.apex.rule.codestyle;
 
 import static net.sourceforge.pmd.properties.PropertyFactory.booleanProperty;
 
+import java.util.HashMap;
+import java.util.Map;
+import java.util.regex.Pattern;
+
 import net.sourceforge.pmd.lang.apex.ast.ASTMethod;
 import net.sourceforge.pmd.lang.apex.ast.ASTProperty;
-import net.sourceforge.pmd.lang.apex.rule.AbstractApexRule;
+import net.sourceforge.pmd.lang.apex.ast.ASTUserEnum;
 import net.sourceforge.pmd.properties.PropertyDescriptor;
 
-public class MethodNamingConventionsRule extends AbstractApexRule {
+public class MethodNamingConventionsRule extends AbstractNamingConventionsRule {
+    private static final Map<String, String> DESCRIPTOR_TO_DISPLAY_NAME = new HashMap<>();
+
+    private static final PropertyDescriptor<Pattern> TEST_REGEX = prop("testPattern", "test method",
+            DESCRIPTOR_TO_DISPLAY_NAME).defaultValue(CAMEL_CASE).build();
+
+    private static final PropertyDescriptor<Pattern> STATIC_REGEX = prop("staticPattern", "static method",
+            DESCRIPTOR_TO_DISPLAY_NAME).defaultValue(CAMEL_CASE).build();
+
+    private static final PropertyDescriptor<Pattern> INSTANCE_REGEX = prop("instancePattern", "instance method",
+            DESCRIPTOR_TO_DISPLAY_NAME).defaultValue(CAMEL_CASE).build();
 
     private static final PropertyDescriptor<Boolean> SKIP_TEST_METHOD_UNDERSCORES_DESCRIPTOR
         = booleanProperty("skipTestMethodUnderscores")
-              .desc("Skip underscores in test methods")
+              .desc("deprecated! Skip underscores in test methods")
               .defaultValue(false)
               .build();
     
     public MethodNamingConventionsRule() {
         definePropertyDescriptor(SKIP_TEST_METHOD_UNDERSCORES_DESCRIPTOR);
+        definePropertyDescriptor(TEST_REGEX);
+        definePropertyDescriptor(STATIC_REGEX);
+        definePropertyDescriptor(INSTANCE_REGEX);
 
         setProperty(CODECLIMATE_CATEGORIES, "Style");
         // Note: x10 as Apex has not automatic refactoring
@@ -35,19 +52,33 @@ public class MethodNamingConventionsRule extends AbstractApexRule {
         if (isOverriddenMethod(node) || isPropertyAccessor(node) || isConstructor(node)) {
             return data;
         }
-        if (isTestMethod(node) && getProperty(SKIP_TEST_METHOD_UNDERSCORES_DESCRIPTOR)) {
+
+        if ("<clinit>".equals(node.getImage()) || "clone".equals(node.getImage())) {
             return data;
         }
 
-        String methodName = node.getImage();
+        if (node.getFirstParentOfType(ASTUserEnum.class) != null) {
+            return data;
+        }
 
-        if (Character.isUpperCase(methodName.charAt(0))) {
-            addViolationWithMessage(data, node, "Method names should not start with capital letters");
+        if (node.getModifiers().isTest()) {
+            if (getProperty(SKIP_TEST_METHOD_UNDERSCORES_DESCRIPTOR)) {
+                checkMatches(TEST_REGEX, CAMEL_CASE_WITH_UNDERSCORES, node, data);
+            } else {
+                checkMatches(TEST_REGEX, node, data);
+            }
+        } else if (node.getModifiers().isStatic()) {
+            checkMatches(STATIC_REGEX, node, data);
+        } else {
+            checkMatches(INSTANCE_REGEX, node, data);
         }
-        if (methodName.indexOf('_') >= 0) {
-            addViolationWithMessage(data, node, "Method names should not contain underscores");
-        }
+
         return data;
+    }
+
+    @Override
+    protected String displayName(String name) {
+        return DESCRIPTOR_TO_DISPLAY_NAME.get(name);
     }
 
     private boolean isOverriddenMethod(ASTMethod node) {
@@ -60,9 +91,5 @@ public class MethodNamingConventionsRule extends AbstractApexRule {
 
     private boolean isConstructor(ASTMethod node) {
         return node.isConstructor();
-    }
-
-    private boolean isTestMethod(ASTMethod node) {
-        return node.getModifiers().isTest();
     }
 }
