+++ pmd-apex/src/test/java/net/sourceforge/pmd/lang/apex/rule/codestyle/LocalVariableNamingConventionsTest.java
@@ -6,6 +6,6 @@ package net.sourceforge.pmd.lang.apex.rule.codestyle;
 
 import net.sourceforge.pmd.testframework.PmdRuleTst;
 
-public class ClassRegexNamingConventionsTest extends PmdRuleTst {
+public class LocalVariableNamingConventionsTest extends PmdRuleTst {
     // no additional unit tests
 }
