--- pmd-apex/src/test/java/net/sourceforge/pmd/lang/apex/rule/codestyle/PropertyRegexNamingConventionsTest.java
@@ -1,11 +1,0 @@
-/**
- * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
- */
-
-package net.sourceforge.pmd.lang.apex.rule.codestyle;
-
-import net.sourceforge.pmd.testframework.PmdRuleTst;
-
-public class PropertyRegexNamingConventionsTest extends PmdRuleTst {
-    // no additional unit tests
-}
