+++ pmd-apex/src/test/resources/net/sourceforge/pmd/lang/apex/rule/codestyle/xml/MethodNamingConventions.xml
@@ -100,4 +100,131 @@ public class Foo {
 }
 		]]></code>
     </test-code>
+
+    <test-code>
+        <description>all is well</description>
+        <expected-problems>0</expected-problems>
+        <code><![CDATA[
+public class Foo {
+    @isTest
+    public static void testMethod() { }
+    public static void staticMethod() { }
+    public void instanceMethod() { }
+}
+        ]]></code>
+    </test-code>
+
+    <test-code>
+        <description>default is camel case</description>
+        <expected-problems>3</expected-problems>
+        <expected-messages>
+            <message>The test method name 'TEST_METHOD' doesn't match '[a-z][a-zA-Z0-9]*'</message>
+            <message>The static method name 'STATIC_METHOD' doesn't match '[a-z][a-zA-Z0-9]*'</message>
+            <message>The instance method name 'INSTANCE_METHOD' doesn't match '[a-z][a-zA-Z0-9]*'</message>
+        </expected-messages>
+        <code><![CDATA[
+public class Foo {
+    @isTest
+    public static void TEST_METHOD() { }
+    public static void STATIC_METHOD() { }
+    public void INSTANCE_METHOD() { }
+}
+        ]]></code>
+    </test-code>
+
+    <test-code>
+        <description>ignores overrides</description>
+        <expected-problems>0</expected-problems>
+        <code><![CDATA[
+public class Foo {
+    public override Integer BAR() { }
+}
+        ]]></code>
+    </test-code>
+
+    <test-code>
+        <description>ignores properties</description>
+        <expected-problems>0</expected-problems>
+        <code><![CDATA[
+public class Foo {
+    public Integer BAR { get; set; }
+}
+        ]]></code>
+    </test-code>
+
+    <test-code>
+        <description>ignores constructors</description>
+        <expected-problems>0</expected-problems>
+        <code><![CDATA[
+public class FOO {
+    public FOO() { }
+}
+        ]]></code>
+    </test-code>
+
+    <test-code>
+        <description>ignores enum methods</description>
+        <rule-property name="instancePattern">Z</rule-property>
+        <expected-problems>0</expected-problems>
+        <code><![CDATA[
+public class Foo {
+    enum Bar {
+    }
+}
+        ]]></code>
+    </test-code>
+
+    <test-code>
+        <description>custom test method pattern</description>
+        <rule-property name="testPattern">[a-zA-Z0-9_]+</rule-property>
+        <expected-problems>2</expected-problems>
+        <expected-messages>
+            <message>The static method name 'STATIC_METHOD' doesn't match '[a-z][a-zA-Z0-9]*'</message>
+            <message>The instance method name 'INSTANCE_METHOD' doesn't match '[a-z][a-zA-Z0-9]*'</message>
+        </expected-messages>
+        <code><![CDATA[
+public class Foo {
+    @isTest
+    public static void TEST_METHOD() { }
+    public static void STATIC_METHOD() { }
+    public void INSTANCE_METHOD() { }
+}
+        ]]></code>
+    </test-code>
+
+    <test-code>
+        <description>custom static method pattern</description>
+        <rule-property name="staticPattern">[a-zA-Z0-9_]+</rule-property>
+        <expected-problems>2</expected-problems>
+        <expected-messages>
+            <message>The test method name 'TEST_METHOD' doesn't match '[a-z][a-zA-Z0-9]*'</message>
+            <message>The instance method name 'INSTANCE_METHOD' doesn't match '[a-z][a-zA-Z0-9]*'</message>
+        </expected-messages>
+        <code><![CDATA[
+public class Foo {
+    @isTest
+    public static void TEST_METHOD() { }
+    public static void STATIC_METHOD() { }
+    public void INSTANCE_METHOD() { }
+}
+        ]]></code>
+    </test-code>
+
+    <test-code>
+        <description>custom instance method pattern</description>
+        <rule-property name="instancePattern">[a-zA-Z0-9_]+</rule-property>
+        <expected-problems>2</expected-problems>
+        <expected-messages>
+            <message>The test method name 'TEST_METHOD' doesn't match '[a-z][a-zA-Z0-9]*'</message>
+            <message>The static method name 'STATIC_METHOD' doesn't match '[a-z][a-zA-Z0-9]*'</message>
+        </expected-messages>
+        <code><![CDATA[
+public class Foo {
+    @isTest
+    public static void TEST_METHOD() { }
+    public static void STATIC_METHOD() { }
+    public void INSTANCE_METHOD() { }
+}
+        ]]></code>
+    </test-code>
 </test-data>
